Use classes to isolate system-specific stuff
